Create types module
